% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coefsChebApprox}
\alias{coefsChebApprox}
\title{Coefficents of Chebyshev approximation by FFT}
\usage{
coefsChebApprox(h_, M, a, b, verbose = 0L)
}
\arguments{
\item{M}{Integer. Order of the polynomial approximation.}

\item{a, b}{Real numbers. Interval on which the approximation is carried out.}

\item{verbose}{Integer. Print messages if \code{verbose=1}.}

\item{h}{R numerical function. Function to be approximated.}

\item{minSubdiv}{Integer. Minimal number of discretisation points.}
}
\value{
A vector containing the \code{M+1} coefficients of the Chebyshev approximation of \code{h} over [\code{a},\code{b}].
}
\description{
Computation of the coefficents of Chebyshev approximation of a function over a given interval (by FFT).
}
\examples{
f<-function(x){
(1+x)**(-2)
}
coefsChebApprox(f,M=100,a=0,b=50)

}
