% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triMetricCoefs.R
\name{coordSph}
\alias{coordSph}
\title{Spherical coordinates}
\usage{
coordSph(x0)
}
\arguments{
\item{x0}{Cartesian coordinates of the point.}
}
\value{
Vector containing the spherical coordinates.
}
\description{
Computes spherical coordinates from Cartesian coordinates.
}
