% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triMetricCoefs.R
\name{.computeTriDiff}
\alias{.computeTriDiff}
\title{Spherical diffusion or metric tensor coefficients}
\usage{
.computeTriDiff(p1, p2, p3, dir, rho1, rho2)
}
\arguments{
\item{p1, p2, p3}{Coordinates of the points defining the triangle.}

\item{dir}{Angle giving the principal direction of diffusion in the tangent space of the sphere.}

\item{rho1}{vector giving the relative weight of the principal diffusion direction.}

\item{rho2}{vector giving the relative weight of the secondary diffusion direction.}
}
\value{
The coefficients of the diffusion or metric tensor for the  triangle.
}
\description{
Compute coefficients of the diffusion or metric tensor on a triangle, for a sphere
}
