% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triMetricCoefs.R
\name{.triMetricPullCoefs}
\alias{.triMetricPullCoefs}
\title{Pullback metric coefficients}
\usage{
.triMetricPullCoefs(p1, p2, p3, Gfunc)
}
\arguments{
\item{p1, p2, p3}{Coordinates of the points defining the triangle.}

\item{Gfunc}{Function returning the 3x3 metric matrix to be pulled back at a given point.}
}
\value{
The coefficients of the matrix of the pulled-back metric on the triangle.
}
\description{
Computes the coefficients of a metric matrix defined by pullback at a given triangle.
}
\keyword{internal}
