% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triMetricCoefs.R
\name{.triMetricSphCoefs}
\alias{.triMetricSphCoefs}
\title{Spherical metric coefficients}
\usage{
.triMetricSphCoefs(p1, p2, p3, GfuncSph)
}
\arguments{
\item{p1, p2, p3}{Coordinates of the points defining the triangle.}

\item{GfuncSph}{Function returning the 2x2 metric matrix in spherical coordinates.}
}
\value{
The coefficients of the metric matrix on the triangle.
}
\description{
Computes the coefficients of a metric matrix defined from spherical coordinates at a given triangle. Only for a UNIT SPHERE.
}
\keyword{internal}
