% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evalCheb}
\alias{evalCheb}
\title{Chebyshev Polynomial}
\usage{
evalCheb(x, n, a = -1, b = 1)
}
\arguments{
\item{x}{A real number or a vector.}

\item{n}{An integer. Order of the Chebyshev polynomial.}

\item{a, b}{Real numbers. Interval on which the Chebyshev polynomial is shifted. Default : \code{a=-1.0}, \code{b=1.0}.}
}
\value{
The evaluation at \code{x} of the \code{n}-th Chebyshev polynomial, shifted to interval \code{[a,b]}.
}
\description{
\code{evalCheb} returns the value of a Chebyshev polynomial evaluated at one (or several) point(s).
}
\details{
This function computes
\eqn{T_n(\frac{2}{b-a}x-\frac{a+b}{b-a})}
where \eqn{T_n} is the n-th Chebyshev poynomial.
}
\examples{
evalCheb(x=runif(10,-2,2), n=10, a=-2, b=2)

}
