% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evalChebDecompMat}
\alias{evalChebDecompMat}
\title{Evaluation of a Chebyshev decomposition}
\usage{
evalChebDecompMat(S, Coefs, a, b)
}
\arguments{
\item{S}{A sparse matrix.}

\item{Coefs}{A numeric vector. Coeffcients of a Chebyshev decomposition.}

\item{a, b}{Real numbers. Interval on which the Chebyshev polynomials are shifted. Default : \code{a=-1.0}, \code{b=1.0}.}
}
\value{
The evaluation at each component at the matrix \code{S} of the Chebyshev decomposition whose coefficients are defined in \code{Coefs}.
}
\description{
\code{evalChebDecompMat} returns the value of a Chebyshev decomposition evaluated at a matrix.
}
\details{
This function computes
\deqn{\sum_{i=0}^{M-1} \mathtt{Coefs}[i+1]T_i(\frac{2}{b-a}S-\frac{a+b}{b-a}I)}
where \code{M=length(Coefs)}, \eqn{I} is the indentity matrix, \eqn{T_i} is the i-th Chebyshev poynomial.
}
