% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evalChebDecompPts}
\alias{evalChebDecompPts}
\title{Evaluation of a Chebyshev decomposition}
\usage{
evalChebDecompPts(x, Coefs, a, b)
}
\arguments{
\item{x}{A numeric vector.}

\item{Coefs}{A numeric vector. Coeffcients of a Chebyshev decomposition.}

\item{a, b}{Real numbers. Interval on which the Chebyshev polynomials are shifted. Default : \code{a=-1.0}, \code{b=1.0}.}
}
\value{
The evaluation at each component of \code{x} of the Chebyshev decomposition whose coefficients are defined in \code{Coefs}.
}
\description{
\code{evalChebDecompPts} returns the value of a Chebyshev decomposition evaluated at several points.
}
\details{
This function computes
\deqn{\sum_{i=0}^{M-1} \mathtt{Coefs}[i+1]T_i(\frac{2}{b-a}x-\frac{a+b}{b-a})}
where \code{M=length(Coefs)}, \eqn{T_i} is the i-th Chebyshev poynomial.
}
\examples{
curve(evalChebDecompPts(x=x,Coefs=runif(5),a=-2,b=2),-2,2)

}
