% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mcsapply}
\alias{mcsapply}
\title{Parallel sapply function}
\usage{
mcsapply(X, FUN, mc.cores = NULL, ...)
}
\arguments{
\item{X}{Vector to which the sapply function if to be applied.}

\item{FUN}{Function to apply to each entry of the vector.}

\item{mc.cores}{Number of cores to use. Default = NULL means that the number of detected cores is used (or the default value in the parallel package if the detection fails).}
}
\value{
Result of the sapply function.
}
\description{
Parallel version of the sapply function, using the mclapply function from the \code{parallel} package.
}
