% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot1d}
\alias{plot1d}
\title{Plot on mesh of 1D surface}
\usage{
plot1d(
  nodeVals,
  nodeMat,
  triMat,
  colorPlot = FALSE,
  r = 0.3,
  col = NULL,
  ttl = "",
  zlim = "range"
)
}
\arguments{
\item{nodeVals}{Vector containing the values, at each mesh node, of the variable to be plotted}

\item{nodeMat}{Coordinates of the nodes defining the segments.}

\item{triMat}{Indices of the nodes forming each segment.}

\item{col}{Vector of colors defining the palette used for the plot (Default=NULL corresponds to a Spectral palette with 192 colors).}

\item{ttl}{Title of the plot.}

\item{zlim}{: Range of values used for color bar: "range" = use the range of values of zplt (Default), "sym" = use a symetric range, or put a vector of size 2 with a custom range.}
}
\value{
Plot of the mesh with colors given by the variable.
}
\description{
2D plot a variable defined on each node of the mesh  of a 1D surface.
}
