% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot1d_deform}
\alias{plot1d_deform}
\title{Plot variable as deformed 1D surface}
\usage{
plot1d_deform(nodeVals, nodeMat, triMat, r = 0.1, ttl = "", ...)
}
\arguments{
\item{nodeVals}{Vector containing the values, at each mesh node, of the variable to be plotted}

\item{nodeMat}{Coordinates of the nodes defining the segments.}

\item{triMat}{Indices of the nodes forming each segment.}

\item{r}{Scaling size for the radial deformation (only used if `colorPlot=FALSE`). The deformed surface, plotted in green, is obtained by translating each node along the normal direction by a length r.}

\item{ttl}{Title of the plot.}

\item{...}{Additional arguments to R base `plot` function}
}
\value{
Plot of the mesh (grey) and of the radially deformed mesh (green).
}
\description{
2D plot of a variable defined on each node of the mesh of a 1D surface, as a deformation of the surface.
DISCLAIMER: Only valid for meshes where the neighbors of a node i are i-1 and i+1.
}
