% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plotOnMesh}
\alias{plotOnMesh}
\title{Plot on mesh of 2D surface}
\usage{
plotOnMesh(zplt, msh, col = NULL, ttl = "", zlim = "range", windowSize = 800)
}
\arguments{
\item{zplt}{Vector containing the values, at each mesh node, of the variable to be plotted}

\item{msh}{Mesh object: Object of class "mesh3d" obtained after importing a mesh through the `vcgImport` function of the `Rvcg` package.}

\item{col}{Vector of colors defining the palette used for the plot (Default=NULL corresponds to a Spectral palette with 192 colors).}

\item{ttl}{Title of the plot.}

\item{zlim}{: Range of values used for color bar: "range" = use the range of values of zplt (Default), "sym" = use a symetric range, or put a vector of size 2 with a custom range.}

\item{windowSize}{: Size of the (square) plotting window.}
}
\value{
Rgl plot of the mesh with colors given by the variable.
}
\description{
3D plot a variable defined on each node of a mesh using Rgl.
}
