% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prodChebDecompVec}
\alias{prodChebDecompVec}
\title{Product of a Chebyshev matrix polynomial and  vectors}
\usage{
prodChebDecompVec(S, Coefs, a, b, V)
}
\arguments{
\item{S}{A sparse matrix.}

\item{Coefs}{A numeric vector. Coeffcients of a Chebyshev decomposition.}

\item{a, b}{Real numbers. Interval on which the Chebyshev polynomials are shifted. Default : \code{a=-1.0}, \code{b=1.0}.}

\item{V}{A matrix.}
}
\value{
A matrix such that each column is the product of the evaluation at the matrix \code{S} of the Chebyshev decomposition, and a column of \code{V}. The coefficients of the decompostion are defined in \code{Coefs}.
}
\description{
\code{prodChebDecompVec} computes the product of a Chebyshev decomposition evaluated at a matrix and vectors binded in a matrix.
}
\details{
This function computes the product
\deqn{\sum_{i=0}^{M-1} \mathtt{Coefs}[i+1]T_i(\frac{2}{b-a}S-\frac{a+b}{b-a}I)V}
where \code{M=length(Coefs)}, \eqn{I} is trhe identity matrix, \eqn{T_i} is the i-th Chebyshev poynomial.
}
