% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPSD.R
\name{simPSD}
\alias{simPSD}
\title{Random field simulation}
\usage{
simPSD(
  psd,
  FEMatList,
  nbsimu = 1,
  seed = NULL,
  WN = NULL,
  tolCoefs = 10^(-4),
  verbose = TRUE
)
}
\arguments{
\item{psd}{Function defining the power spectral density of the field.}

\item{FEMatList}{List of FEM matrices obtained from the \code{matFEM} function.}

\item{nbsimu}{Number of samples to generate (Default=1).}

\item{seed}{Seed for random number generator (Default=NULL).}

\item{WN}{Matrix whose columns contains vectors of Gaussian white noise to be used to generate the samples. If set to NULL (Default), this matrices is generated on the fly. If set, \code{nbsimu} and \code{seed} are ignored. Note: The number of rows should be equal to the size of the FEM matrices.}

\item{tolCoefs}{Tolerance to discard Chebyshev coefficients (Default=10^{-4}). The coefficients below tolCoefs*maximum coeffients are discarded.}

\item{verbose}{Boolean. Print messages or not.}
}
\value{
A matrix containing the generated samples.
}
\description{
Generate samples of a Gaussian random field specified by a power spectral density and FEM matrices.
}
