% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triMetricCoefs.R
\name{triMetricFromValues}
\alias{triMetricFromValues}
\title{Spherical diffusion or metric tensor coefficients}
\usage{
triMetricFromValues(nodeMat, triMat, dirNode, rhoNode)
}
\arguments{
\item{nodeMat}{Coordinates of the points defining the triangle.}

\item{triMat}{Indices of the nodes forming each triangle.}

\item{dirNode}{Angle giving the principal direction of diffusion in the tangent space of each node of the sphere.}

\item{rhoNode}{Matrix giving the relative weights of the diffusion directions at each node.}
}
\value{
The coefficients of the diffusion or metric tensor for each triangle.
}
\description{
Compute coefficients of the diffusion or metric tensor on each triangle of a mesh, for a sphere
}
