% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triMetricCoefs.R
\name{triMetricPull1d}
\alias{triMetricPull1d}
\title{Pullback metric coefficients for 1D surface}
\usage{
triMetricPull1d(
  nodeMat,
  triMat,
  Gfunc = NULL,
  useParallel = TRUE,
  mc.cores = NULL
)
}
\arguments{
\item{nodeMat}{Coordinates of the nodes defining the segments.}

\item{triMat}{Indices of the nodes forming each segment.}

\item{Gfunc}{Function returning the metric matrix to be pulled back at a given point. If NULL, the Euclidean metric is used.}

\item{useParallel}{Whether to parallelize the computations using the `parallel` package. Default = TRUE.}

\item{mc.cores}{Number of cores to use when using parallel computations. Default = NULL means that the number of detected cores is used (or the default value in the parallel package if the detection fails).}
}
\value{
The coefficients of the matrix of the pulled-back metric for each triangle.
}
\description{
Computes the coefficients of a metric matrix defined by pullback for each segment of the discretization of a 1D surface.
}
