% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triMetricCoefs.R
\name{triMetricSph}
\alias{triMetricSph}
\title{Spherical metric coefficients}
\usage{
triMetricSph(
  nodeMat,
  triMat,
  GfuncSph = NULL,
  useParallel = TRUE,
  mc.cores = NULL
)
}
\arguments{
\item{nodeMat}{Coordinates of the points defining the triangle.}

\item{triMat}{Indices of the nodes forming each triangle.}

\item{GfuncSph}{Function returning the 2x2 metric matrix in spherical coordinates. If NULL, the canonical metric on the sphere is used.}

\item{useParallel}{Whether to parallelize the computations using the `parallel` package. Default = TRUE.}

\item{mc.cores}{Number of cores to use when using parallel computations. Default = NULL means that the number of detected cores is used (or the default value in the parallel package if the detection fails).}
}
\value{
The coefficients of the metric matrix for each triangle.
}
\description{
Computes the coefficients of a metric matrix defined from spherical coordinates for each triangle of a triangulation. Only for a UNIT SPHERE.
}
