% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triPotCoefs.R
\name{triPot}
\alias{triPot}
\title{Potential function coefficients}
\usage{
triPot(
  nodeMat,
  triMat,
  potf = 1,
  coordSph = FALSE,
  useParallel = TRUE,
  mc.cores = NULL
)
}
\arguments{
\item{nodeMat}{Coordinates of the points defining the triangle.}

\item{triMat}{Indices of the nodes forming each triangle.}

\item{potf}{Potential function. If set to a (positive) constant, the corresponding constant function is used. Otherwise, it should be a positive function of the coordinates.}

\item{coordSph}{Logical, whether the potential function is defined as function of the spherical coordinates (Default=FALSE).}

\item{useParallel}{Whether to parallelize the computations using the `parallel` package. Default = TRUE.}

\item{mc.cores}{Number of cores to use when using parallel computations. Default = NULL means that the number of detected cores is used (or the default value in the parallel package if the detection fails).}
}
\value{
The coefficients of the potential function for each triangle.
}
\description{
Computes the coefficients of a potential function.
}
